add KafkaMessageIterator interface to isolate kafka lib usage
the main idea is to isolate kafka interaction so we're able to introduce a new profile with the new
kafka consumer
