+++ deploys/helm/secor/config/secor.kubernetes-test.properties
@@ -207,6 +207,12 @@ kafka.useTimestamp=false
 # `com.pinterest.secor.timestamp.Kafka10MessageTimestamp`.
 kafka.message.timestamp.className=com.pinterest.secor.timestamp.Kafka10MessageTimestamp
 
+# Classname for the message iterator you want to use. The message iterator determines what kind of consumer
+# secor will use to communicate with kafka. com.pinterest.secor.reader.LegacyKafkaMessageIterator uses
+# the old kafka consumer written scala. Its not recommended to use the legacy iterator with kafka version >= 1.0 since it
+# does not support the new broker protocols. You may face significant peformance degration on your brokers if you use it
+kafka.message.iterator.className=com.pinterest.secor.reader.LegacyKafkaMessageIterator
+
 # Secor generation is a version that should be incremented during non-backwards-compatible
 # Secor releases. Generation number is one of the components of generated log file names.
 # Generation number makes sure that outputs of different Secor versions are isolated.
