+++ src/main/java/com/pinterest/secor/consumer/Consumer.java
@@ -29,7 +29,6 @@ import com.pinterest.secor.uploader.UploadManager;
 import com.pinterest.secor.uploader.Uploader;
 import com.pinterest.secor.util.ReflectionUtil;
 import com.pinterest.secor.writer.MessageWriter;
-import kafka.consumer.ConsumerTimeoutException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -120,18 +119,11 @@ public class Consumer extends Thread {
 
     // @return whether there are more messages left to consume
     protected boolean consumeNextMessage() {
-        Message rawMessage = null;
-        try {
-            boolean hasNext = mMessageReader.hasNext();
-            if (!hasNext) {
-                return false;
-            }
-            rawMessage = mMessageReader.read();
-        } catch (ConsumerTimeoutException e) {
-            // We wait for a new message with a timeout to periodically apply the upload policy
-            // even if no messages are delivered.
-            LOG.trace("Consumer timed out", e);
+        boolean hasNext = mMessageReader.hasNext();
+        if (!hasNext) {
+            return false;
         }
+        Message rawMessage = mMessageReader.read();
         if (rawMessage != null) {
             // Before parsing, update the offset and remove any redundant data
             try {
